Feat: default 전체 폴더 추가, FolderList 컴포넌트 분리
