+++ src/app/collection/page.tsx
@@ -1,24 +1,22 @@
 'use client';
 
-import Link from 'next/link';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { ChangeEvent, useState } from 'react';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { isAxiosError } from 'axios';
 
 import * as styles from './page.css';
 
 import Header from '@/components/Header/Header';
 import BottomSheet from '@/components/BottomSheet/ver3.0/BottomSheet';
-import FolderIcon from '@/components/icons/FolderIcon';
+import FolderList from './_components/FolderList';
 
 import useBooleanOutput from '@/hooks/useBooleanOutput';
 import { useLanguage } from '@/store/useLanguage';
 import { useUser } from '@/store/useUser';
 
 import createCollectionFolder from '../_api/folder/createFolder';
-import getFolders, { FoldersResponseType } from '../_api/folder/getFolders';
 
 import toasting from '@/lib/utils/toasting';
 import toastMessage from '@/lib/constants/toastMessage';
@@ -29,11 +27,6 @@ export default function CollectionPage() {
   const { user: userMe } = useUser();
   const queryClient = useQueryClient();
   const { language } = useLanguage();
-  const { data } = useQuery<FoldersResponseType>({
-    queryKey: [QUERY_KEYS.getFolders],
-    queryFn: getFolders,
-    staleTime: 1000 * 60 * 5, // 5분 설정
-  });
 
   const { isOn, handleSetOn, handleSetOff } = useBooleanOutput(false);
   const [value, setValue] = useState('');
@@ -79,17 +72,7 @@ export default function CollectionPage() {
     <section className={styles.wrapper}>
       <Header title="콜렉션" left="back" leftClick={() => router.push(`/user/${userMe.id}/mylist`)} />
       <div className={styles.container}>
-        <div className={styles.folders}>
-          {data?.folders.map((folder) => (
-            <Link href={`/collection/${folder.folderId}`} key={folder.folderId} className={styles.folder}>
-              <FolderIcon />
-              <p className={styles.title}>
-                <span className={styles.folderName}>{folder.folderName}</span>
-                {folder.listCount > 0 && <span>{`(${folder.listCount})`}</span>}
-              </p>
-            </Link>
-          ))}
-        </div>
+        <FolderList />
         <div className={styles.addFolderButtonContainer}>
           <button className={styles.addFolderButton} onClick={handleSetOn}>
             <Image src={'/icons/new/add.svg'} width={16} height={16} alt="폴더 추가하기" />
